rip out custom component testing framework
