+++ src/pages/Home/Home.html
@@ -1,5 +1,9 @@
 <main id=main class='main home-page' data-page=Home>
 
+  <style>
+    {{> home-css}}
+  </style>
+
   <h1>Lotus</h1>
 
   <p>Coming soon!</p>
