--- src/test.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang=en>
-<head>
-    <meta charset=utf-8>
-    <title>Lotus Test Page</title>
-</head>
-<body>
-
-</body>
-</html>
